Implemented up to 4 multiedges
The edges are added and removed correctly. Every edge now has a list of siblings (which is being maintained by the Structure class that adds the edges)

createEdge needs to be removed asap. Creating an edge changes the structure, even when not adding it (by connecting to exisiting vertices)

Directional edge endpoint needs to be correctly calculated depending on the edge's offset.