Implemented dashed and dotted edges
Edges now have a LineType (either plain, dotted or dashed. See gralog.rendering.GralogGraphicsContext.LineType)

The line type is now a parameter for drawing lines or loops in the GraphicsContext classes. Also, the ReflectedView now captures line type
and correctly rewrites the linetype field of the edge objects. It displays a ChoiceBox in the inspector.