More Basic Functionality, Started Graph Export
Added functionality to get all nodes, all edges, etc. and started working on functionality to get a graph as a string. The problem with this is that they all have newlines in them so i will have to come up with a cleverer solution that will involve programmers who don't use the library to have a greater knowledge of this functionality. That is ok, but i/o tends to lead to errors...