Implemented intersection calculation for rectangle
Rectangles now correctly calculate the point of intersection between an external point a and an internal point b. The intersection can be used to calculate the correct positioning of the edge endpoint.

Now that that is implemented consider adding additional spacing options either globally (structure) or locally (single vertex, single edge)