Added basic FXML Preference window
Preference window gets loaded from FXML file. Style and xml files get loaded from resource of class loader.

All control nodes in the preference window can have an ID so that if their ID matches a preference variable descriptor of a
member of a class in gralogfx, the output of the control will be formatted and written to the preference file with the correct key.

Annotations for preference variable descriptors need to be implemented