Started working on loading relevant prefs into the Pref controls
Problem: There is no guarantee that certain global preferences are already in the preference file.
Make sure that when creating a preference file, all relevant fields that are a @PreferenceVariable
get a default entry in the config file.

Also need to do exception handling for when preferences entries aren't set and the code gets called
from a unit test.