More edge id's bug fixes
I fixed a bunch of bugs pertaining to edge id's, now all the edge commands can optionally be performed with an id. Added a command to delete all edges. I have decided, however, that the edge id's should be global (like the vertices) rather than local. It's just annoying otherwise. For example, if you wanted to delete the edge with id = x, you'd also have to know exactly which vertices it lies between before doing so, which is occaisionally quite impratical.