Finished Cubic Bezier box-selection
Cubic bezier curves can now be selected by a box selection. Relevant functionality is found in BezierUtilities and Structure.findObjects()

A bezier curve is selected when an x-intersection between box and curve is between the y-bounds of the box, a y-intersection is between the x-bounds
of the box or when one of the curve starting points (c0 and c3/c2) is contained by the box.

Note: consider adding a rough geometric filter (much like the seperating axis test) so we don't have to calculate the intersections