Algorithms/generators fixed up
After all the work I did this morning, changing the depencies seems to have fixed the crux of the problem. Thanks @adrian lol and I feel stupid for spending so much time not really solving the problem but that's life. Still some minor tweaks to some of the generators whose vertices are not shown