Finished deleting console logs, other general debugging
Now that the prints are gone everything is much faster; this has led to a concurrency error that I had never seen before. I made a quick and dirty fix that I think is acceptable (transfer the collection to an external list, iterate through new list rather than collection that is being modified).