Fix broken unit tests.
The unit tests had been broken for a while.

After this change, "./gradlew test" should work again.

Some unit tests are harder to fix because they try to open dialogs,
which is inherently incompatible with units tests because they need to
run fully automatically on the continuous integration system
"Travis" (and also with "./gradlew test").

This works towards fixing the Travis build.  The Travis build is not
expected to succeed after this change because "./gradlew check" still
fails due to new code that is incompatible with the declared coding
style.